fix: correct nine correctness bugs found by full-code audit (#59)
- diffusion/schedulers: FlowMatchEulerDiscreteScheduler sigmas now descend
  1 -> 0 with terminal 0.0 (diffusers parity); previously they ascended,
  integrating clean -> noise with a no-op final step. step() past the end
  now raises instead of silently wrapping.
- diffusion/ddim: _prev_timestep looks up the actual schedule instead of
  recomputing (off-by-one vs the rounded schedule); out-of-schedule
  timesteps and num_inference_steps > num_train_timesteps now raise.
- conv/weight_norm: stop writing module.__dict__, which permanently
  shadowed the wrapped weight and broke update()/load_weights(); weight
  swap is now exception-safe (try/finally).
- tiling/spatial: per-axis output scale computed from the actual slice
  length; non-square inputs with one axis smaller than tile_size no
  longer crash. tiling/temporal: overlap >= window_size now raises.
- encoding/fourier: frequencies stored as a non-parameter buffer; it was
  a trainable parameter, breaking optimizers and strict weight loading.
- diffusion/cfg_skip + attention_cache: similarity/summary reductions run
  in float32; fp16 accumulation overflowed to nan for realistic sizes.
- rope: apply_rotary_emb gains seq_axis=; ambiguous sequence-axis
  detection (batch == seq) now raises instead of rotating the batch axis.
- rasterize: depth_prior shape validated, closing a silent GPU OOB read.
- attention/masks, moe, diffusion/teacache: argument validation (window
  size, top_k range, num_steps/threshold/step_index) matching sibling
  modules.

Every fix is covered by a regression test written first (red) and
verified green; full suite 498 passed, ruff + ty clean.

+++ src/mlx_arsenal/diffusion/cfg_skip.py
@@ -65,8 +65,10 @@ def cfg_head_similarity(
         )
     num_heads = cond.shape[1]
     # Flatten to (num_heads, K) where K = B * S * D (and any trailing axes).
-    cond_h = mx.swapaxes(cond, 0, 1).reshape(num_heads, -1)
-    uncond_h = mx.swapaxes(uncond, 0, 1).reshape(num_heads, -1)
+    # Reduce in float32: fp16 accumulation overflows for realistic
+    # activation magnitudes (sum of B*S*D values).
+    cond_h = mx.swapaxes(cond, 0, 1).reshape(num_heads, -1).astype(mx.float32)
+    uncond_h = mx.swapaxes(uncond, 0, 1).reshape(num_heads, -1).astype(mx.float32)
     if metric == "cosine":
         dot = mx.sum(cond_h * uncond_h, axis=1)
         c_norm = mx.sqrt(mx.sum(cond_h * cond_h, axis=1))
